@@ -30,132 +30,46 @@
         ]
     },
     "devDependencies": {
-        "@babel/core": "7.12.3",
-        "@babel/preset-env": "^7.14.7",
-        "@pmmmwh/react-refresh-webpack-plugin": "0.4.3",
-        "@svgr/webpack": "5.5.0",
-        "@testing-library/jest-dom": "^5.11.4",
-        "@testing-library/react": "^11.1.0",
-        "@testing-library/user-event": "^12.1.10",
-        "@types/jest": "^26.0.15",
-        "@types/node": "^12.0.0",
-        "@types/react": "^17.0.0",
-        "@types/react-dom": "^17.0.0",
-        "@typescript-eslint/eslint-plugin": "^4.5.0",
+        "@babel/core": "7.18.9",
+        "@babel/preset-env": "^7.18.9",
+        "@babel/preset-react": "^7.18.6",
+        "@babel/preset-typescript": "^7.18.6",
+        "@testing-library/jest-dom": "^5.16.4",
+        "@testing-library/react": "^13.3.0",
+        "@types/jest": "^28.1.6",
+        "@types/node": "^18.0.6",
+        "@types/react": "^18.0.15",
+        "@types/react-dom": "^18.0.6",
+        "@types/testing-library__jest-dom": "^5.14.5",
+        "@typescript-eslint/eslint-plugin": "^5.30.0",
         "@typescript-eslint/parser": "^4.5.0",
         "apca-w3": "^0.1.1",
-        "babel-eslint": "^10.1.0",
-        "babel-jest": "^26.6.0",
-        "babel-loader": "8.1.0",
-        "babel-plugin-named-asset-import": "^0.3.7",
-        "babel-preset-react-app": "^10.0.0",
-        "bfj": "^7.0.2",
-        "camelcase": "^6.1.0",
-        "case-sensitive-paths-webpack-plugin": "2.3.0",
-        "css-loader": "4.3.0",
-        "dotenv": "8.2.0",
-        "dotenv-expand": "5.1.0",
-        "eslint": "^7.11.0",
+        "babel-loader": "^8.2.5",
+        "css-loader": "^6.7.1",
+        "eslint": "^8.20.0",
         "eslint-config-react-app": "^6.0.0",
-        "eslint-plugin-flowtype": "^5.2.0",
-        "eslint-plugin-import": "^2.22.1",
-        "eslint-plugin-jest": "^24.1.0",
-        "eslint-plugin-jsx-a11y": "^6.3.1",
-        "eslint-plugin-react": "^7.21.5",
-        "eslint-plugin-react-hooks": "^4.2.0",
-        "eslint-plugin-testing-library": "^3.9.2",
-        "eslint-webpack-plugin": "^2.5.2",
-        "file-loader": "6.1.1",
-        "fs-extra": "^9.0.1",
-        "html-webpack-plugin": "4.5.0",
-        "identity-obj-proxy": "3.0.0",
-        "jest": "26.6.0",
-        "jest-circus": "26.6.0",
-        "jest-resolve": "26.6.0",
-        "jest-watch-typeahead": "0.6.1",
-        "mini-css-extract-plugin": "0.11.3",
-        "pnp-webpack-plugin": "1.6.4",
-        "prettier": "^2.3.2",
-        "prompts": "2.4.0",
+        "eslint-plugin-import": "^2.26.0",
+        "eslint-plugin-jsx-a11y": "^6.6.0",
+        "eslint-plugin-react": "^7.30.0",
+        "eslint-plugin-react-hooks": "^4.6.0",
+        "eslint-plugin-testing-library": "^5.5.1",
+        "html-webpack-plugin": "^5.5.0",
+        "identity-obj-proxy": "^3.0.0",
+        "jest": "28.1.3",
+        "jest-environment-jsdom": "^28.1.3",
+        "jest-watch-typeahead": "2.0.0",
+        "prettier": "^2.7.1",
         "protean-elements": "^0.3.0",
         "protean-styles": "^0.3.0",
-        "react": "^17.0.2",
-        "react-app-polyfill": "^2.0.0",
-        "react-dev-utils": "^11.0.3",
-        "react-dom": "^17.0.2",
-        "react-refresh": "^0.8.3",
-        "resolve": "1.18.1",
-        "resolve-url-loader": "^3.1.2",
-        "sass": "^1.37.5",
-        "sass-loader": "^10.0.5",
-        "semver": "7.3.2",
-        "style-loader": "1.3.0",
-        "terser-webpack-plugin": "4.2.3",
-        "ts-pnp": "1.2.0",
-        "typescript": "^4.1.2",
-        "url-loader": "4.1.1",
-        "web-vitals": "^1.0.1",
-        "webpack": "4.44.2",
-        "webpack-dev-server": "3.11.1",
-        "webpack-manifest-plugin": "2.2.0",
-        "workbox-webpack-plugin": "5.1.4"
+        "react": "^18.2.0",
+        "react-dom": "^18.2.0",
+        "sass": "^1.53.0",
+        "sass-loader": "^13.0.2",
+        "style-loader": "^3.3.1",
+        "typescript": "^4.7.4",
+        "webpack": "^5.73.0",
+        "webpack-cli": "^4.10.0",
+        "webpack-dev-server": "^4.9.3"
     },
-    "jest": {
-        "roots": [
-            "<rootDir>/src"
-        ],
-        "collectCoverageFrom": [
-            "src/**/*.{js,jsx,ts,tsx}",
-            "!src/**/*.d.ts"
-        ],
-        "setupFiles": [
-            "react-app-polyfill/jsdom"
-        ],
-        "setupFilesAfterEnv": [
-            "<rootDir>/src/setupTests.ts"
-        ],
-        "testMatch": [
-            "<rootDir>/src/**/__tests__/**/*.{js,jsx,ts,tsx}",
-            "<rootDir>/src/**/*.{spec,test}.{js, jsx,ts,tsx}"
-        ],
-        "testEnvironment": "jsdom",
-        "testRunner": "/Users/abdulmhamid/proteanstudio/contrast-demo/node_modules/jest-circus/runner.js",
-        "transform": {
-            "^.+\\.(js|jsx|mjs|cjs|ts|tsx)$": "<rootDir>/config/jest/babelTransform.js",
-            "^.+\\.css$": "<rootDir>/config/jest/cssTransform.js",
-            "^(?!.*\\.(js|jsx|mjs|cjs|ts|tsx|css|json)$)": "<rootDir>/config/jest/fileTransform.js"
-        },
-        "transformIgnorePatterns": [
-            "[/\\\\]node_modules[/\\\\].+\\.(js|jsx|mjs|cjs|ts|tsx)$",
-            "^.+\\.module\\.(css|sass|scss)$"
-        ],
-        "modulePaths": [],
-        "moduleNameMapper": {
-            "^react-native$": "react-native-web",
-            "^.+\\.module\\.(css|sass|scss)$": "identity-obj-proxy"
-        },
-        "moduleFileExtensions": [
-            "web.js",
-            "js",
-            "web.ts",
-            "ts",
-            "web.tsx",
-            "tsx",
-            "json",
-            "web.jsx",
-            "jsx",
-            "node"
-        ],
-        "watchPlugins": [
-            "jest-watch-typeahead/filename",
-            "jest-watch-typeahead/testname"
-        ],
-        "resetMocks": true
-    },
-    "babel": {
-        "presets": [
-            "@babel/preset-env",
-            "react-app"
-        ]
-    }
+    "packageManager": "yarn@3.2.1"
 }
