@@ -1,19 +1,15 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
+import { StrictMode } from 'react';
+import { createRoot } from 'react-dom/client';
 import App from './components/App';
-// import reportWebVitals from './reportWebVitals';
 import { defineCustomElements } from 'protean-elements/loader';
 
 defineCustomElements();
 
-ReactDOM.render(
-    <React.StrictMode>
-        <App />
-    </React.StrictMode>,
-    document.getElementById('root')
-);
+const rootElement = document.querySelector<HTMLDivElement>('#root');
+const root = createRoot(rootElement!);
 
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-// reportWebVitals(console.log);
+root.render(
+    <StrictMode>
+        <App />
+    </StrictMode>
+);
